Se muestra toast de la bateria
Se agrego al inicio de la aplicacion, un toast indicando el porcentaje de la bateria.

+++ app/src/main/java/com/example/TP2/usecase/GetBatteryPercentage.java
@@ -13,8 +13,8 @@ public class GetBatteryPercentage {
         int level = batteryStatus.getIntExtra(BatteryManager.EXTRA_LEVEL, -1);
         int scale = batteryStatus.getIntExtra(BatteryManager.EXTRA_SCALE, -1);
 
-        float batteryPct = level * 100 / (float)scale;
+        int batteryPct = level * 100 / scale;
 
-        return Float.toString(batteryPct);
+        return Integer.toString(batteryPct);
     }
 }
